move the boundary building into another repository
the boundary building is needed either by mimir and by cosmogony, so we
move it into another crate.
