--- libs/osm_builder/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "osm_builder"
-version = "1.2.0"
-authors = ["dt.ro <dt.ro@canaltp.fr>"]
-
-[dependencies]
-osmpbfreader = "*"
-
-[dependencies.mimir]
-path = "../mimir"
